[CLOSES #637] Split up Topbar into feature with many sub components
