+++ misk/web/@misk/components/src/components/index.ts
@@ -3,5 +3,3 @@ export * from "./OfflineComponent"
 export * from "./PathDebugComponent"
 export * from "./SidebarComponent"
 export * from "./TextHTMLOrElementComponent"
-export * from "./TopbarComponent"
-export * from "./TopbarComponentDimensionAware"
