+++ misk/web/@misk/components/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./builders"
 export * from "./components"
 export * from "./containers"
+export * from "./features"
+export * from "./utilities"
